+++ source_files/edge/r_things.cc
@@ -387,8 +387,8 @@ static void RGL_DrawPSprite(pspdef_t * psp, int which,
 		GLuint fuzz_tex = is_fuzzy ? W_ImageCache(fuzz_image, false) : 0;
 
 		local_gl_vert_t * glvert = RGL_BeginUnit(GL_POLYGON, 4,
-				 is_additive ? ENV_SKIP_RGB : GL_MODULATE, tex_id,
-				 is_fuzzy ? GL_MODULATE : ENV_NONE, fuzz_tex,
+				 is_additive ? (GLuint) ENV_SKIP_RGB : GL_MODULATE, tex_id,
+				 is_fuzzy ? GL_MODULATE : (GLuint) ENV_NONE, fuzz_tex,
 				 pass, blending, pass > 0 ? RGB_NO_VALUE : fc_to_use,
 				 fd_to_use);
 
@@ -1518,8 +1518,8 @@ void RGL_DrawThing(drawfloor_t *dfloor, drawthing_t *dthing)
 		GLuint fuzz_tex = is_fuzzy ? W_ImageCache(fuzz_image, false) : 0;
 
 		local_gl_vert_t * glvert = RGL_BeginUnit(GL_POLYGON, 4,
-				 is_additive ? ENV_SKIP_RGB : GL_MODULATE, tex_id,
-				 is_fuzzy ? GL_MODULATE : ENV_NONE, fuzz_tex,
+				 is_additive ? (GLuint) ENV_SKIP_RGB : GL_MODULATE, tex_id,
+				 is_fuzzy ? GL_MODULATE : (GLuint) ENV_NONE, fuzz_tex,
 				 pass, blending, pass > 0 ? RGB_NO_VALUE : fc_to_use,
 				 fd_to_use);
